backend/requirements.txt
python-telegram-bot
twilio
numba
moviepy>=2.0.0
numpy
pymupdf
websockets
reportlab
fastapi
uvicorn
python-multipart
pydantic
firebase-admin
python-dotenv
google-generativeai
groq
cloudinary
requests
google-cloud-documentai
jinja2
qrcode[pil]
Pillow
